requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
pytest==7.4.0
pandas==2.0.3
python-docx==0.8.11
PyPDF2==3.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.10
langchain-community==0.0.10
